requirements.txt
loguru~=0.7.2
httpx~=0.27.0
PyYAML~=6.0.1
cachetools~=5.3.3